requirements/base.txt
Django==4.0.7
django-environ==0.9.0
djangorestframework==3.13.1
psycopg2-binary==2.9.5


whitenoise==6.2.0

django-filter==22.1
django-extensions==3.2.1
django-cors-headers==3.13.0
django-storages==1.13.1

drf-jwt==1.19.2

boto3==1.24.71
attrs==22.1.0

djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2

django-redis==5.2.0
Faker==15.1.1
factory-boy==3.2.1
pytest==7.2.0
pytest-django==4.5.2
